Update PIDX reader: domain decomposition fixes and CMake
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@31349 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/databases/IDX/pidx_idx_io.cpp
@@ -17,10 +17,10 @@
 
 #include <InvalidFilesException.h>
 #include <DebugStream.h>
-
 #include <cstdarg>
 #include <string>
-
+#include "pidx_idx_io.h"
+#include "PIDX.h"
 //#include "data_handle/PIDX_data_types.h"
 
 static PIDX_point global_size, local_offset, local_size;
@@ -339,14 +339,14 @@ unsigned char* PIDXIO::getData(const VisitIDXIO::Box box, const int timestate, c
   int v_per_sample = 0;
   PIDX_values_per_datatype(variable->type_name, &v_per_sample, &bits_per_sample);
 
-  int this_size = (int)(local_size[0] * local_size[1] * local_size[2]);
+  uint32_t this_size = (uint32_t)(local_size[0] * local_size[1] * local_size[2]);
 
-  void *data = malloc((bits_per_sample/8) * this_size * v_per_sample);//variable->values_per_sample);
-  memset(data, 0, (bits_per_sample/8) * this_size * v_per_sample);//variable->values_per_sample);
+  void *data = malloc((uint32_t)((bits_per_sample/8) * this_size * v_per_sample));//variable->values_per_sample);
+  memset(data, 0, ((uint32_t)(bits_per_sample/8) * this_size * v_per_sample));//variable->values_per_sample);
 
   ret = PIDX_variable_read_data_layout(variable, local_offset, local_size, data, PIDX_row_major);
   if (ret != PIDX_success)  terminate_with_error_msg("PIDX_variable_read_data_layout");
-  
+
   ret = PIDX_close(pidx_file);
   if (ret != PIDX_success)  terminate_with_error_msg("PIDX_close");
